fix multilabel loss evaluate

+++ src/learners/classificator_learner.py
@@ -36,7 +36,7 @@ class ClassificatorLearner(LightningModule):
     def training_step(self, batch, batch_idx):
         x, y = batch
         y_ = self.forward(x)
-        loss = self.loss_f(y_, y if self.cfg.loss.multilabel else y.argmax(dim=1))
+        loss = self.loss_f(y_, y.float() if self.cfg.loss.multilabel else y.argmax(dim=1))
         self.log('train/loss', loss, on_step=True, on_epoch=False)
         self.train_metrics.update(y_, y)
         return loss
@@ -50,7 +50,7 @@ class ClassificatorLearner(LightningModule):
     def validation_step(self, batch, batch_idx):
         x, y = batch
         y_ = self.forward(x)
-        loss = self.loss_f(y_, y if self.cfg.loss.multilabel else y.argmax(dim=1))
+        loss = self.loss_f(y_, y.float() if self.cfg.loss.multilabel else y.argmax(dim=1))
         self.val_metrics.update(y_, y)
         self.log(f'val/loss', loss, on_step=False, on_epoch=True)
         return loss
